Project Completed

+++ phishing/components/model_evaluation.py
@@ -73,7 +73,7 @@ class ModelEvaluation:
             # accuracy using previously trained model
             input_feature_name = list(transformer.feature_names_in_)
             input_arr = transformer.transform(test_df[input_feature_name])
-            y_pred = model.predict(input_arr)
+            y_pred = model.predict(input_arr).astype('int')
             print(f"Prediction Using Previous Model : {target_encoder.inverse_transform(y_pred[:5])}")
             previous_model_score = f1_score(y_true = y_true,y_pred=y_pred)
             logging.info(f"Accuracy Using Previous trained Model :{previous_model_score}")
@@ -81,7 +81,7 @@ class ModelEvaluation:
             #accuracy using current trained model
             input_feature_name = list(transformer.feature_names_in_)
             input_arr = current_transformer.transform(test_df[input_feature_name])
-            y_pred = current_model.predict(input_arr)
+            y_pred = current_model.predict(input_arr).astype('int')
             y_true = current_target_encoder.transform(target_df)
             print(f"Prediction Using trained Model : {current_target_encoder.inverse_transform(y_pred[:5])}")
             current_model_score = f1_score(y_true = y_true,y_pred=y_pred)
